Add files via upload
Snake is a classic game where the player controls a snake moving on the screen. The goal is to eat food that appears randomly, causing the snake to grow longer. As the snake grows, it becomes more challenging to maneuver. The player must avoid collisions with the screen's edges or the snake's own body.